GCPC/requirements.txt
numpy>=1.24
onnxruntime-directml>=1.17; sys_platform == 'win32'
opencv-python>=4.8
